Propagate HoleFilledEvent from Ring to Rings to Mycelium
Delete Perimeter. Mycelium can use Rings directly without Perimeter.

+++ src/app/Settings.h
@@ -46,7 +46,7 @@ struct BorderSettings {
 };
 
 struct HolesSettings {
-  int space;
+  float space;
   int resolution;
   int maxRings;
 };
